+++ app/src/main/java/com/ebata_shota/holdemstacktracker/domain/model/PodState.kt
@@ -3,7 +3,7 @@ package com.ebata_shota.holdemstacktracker.domain.model
 data class PodState(
     val id: Long,
     val podNumber: Int,
-    val podSize: Float,
+    val podSize: Double,
     val involvedPlayerIds: List<PlayerId>,
     val isClosed: Boolean
 )
