+++ app/src/main/java/com/ebata_shota/holdemstacktracker/domain/usecase/GetMaxBetSizeUseCase.kt
@@ -3,5 +3,5 @@ package com.ebata_shota.holdemstacktracker.domain.usecase
 import com.ebata_shota.holdemstacktracker.domain.model.BetPhaseActionState
 
 interface GetMaxBetSizeUseCase {
-    fun invoke(actionStateList: List<BetPhaseActionState>): Float
+    fun invoke(actionStateList: List<BetPhaseActionState>): Double
 }
